Add support for x2apic
* Since not all PC's do not use XAPIC and some use X2APIC instead (like
my pc), we have to support X2APIC to be able to boot the OS on real
hardware. This commit adds support for X2APIC and has been tested on
real hardware and works perfectly fine and boots the APs whihout any
issues. There is an issue that I am looking into that it gets STUCK
after loading the APs ONLY ON REAL HARDWARE. I am trying to find what
the possible problem could be as its not even triple faulting and no
panic what so ever. And I cannot just hook an arduino to my PC and see
the registers and know what is going wrong.