aero_build: rewrite it in python :^)
This will allow us for extra flexibility, because Python runs very
quick, compared to previous aero_build, which had to compile dozen
of crates before being able to do anything.

This change also starts the transition over to ISO-based distributions
of aero, which will allow use of a single disk image for both legacy
and UEFI targets as well as support for all major storage medium, which
means a single image for optical disks, hard disks and removable disks.