sys_getdents: read one directory entry at a time
Unfortunately, the sys_read_entries sysdep in libc is required to read one
directory entry at a time but the Aero getdents syscall reads as much it can
into the provided in the buffer and we dont know the size of the buffer
since we dont know the size of the name aswell. We cannot also assign
an aribiray size to the name since the syscall will bail out if the size
of the buffer is too small to fit in the name. :^)

* We need a way to retrieve the chopped off name.
* We need to null terminate the string since rust strings are not null
terminated and this could possibally cause issues.

Signed-off-by: Andy-Python-Programmer <andypythonappdeveloper@gmail.com>