ipc: initial implementation
This patch implements a basic IPC system :^) There is still a lot
of work to do, but it's something we can work upon. The current system
works in following way:

There are 2 system calls available:
 * sys_ipc_send(target_pid, *buffer, flags);
 * sys_ipc_recv(*buffer, *sender, *length, flags);

Both of these system calls return the message ID, so the sender
can wait for a response to that message and the receipient can
attach some sort of information that will indicate to which message
they are replying, that could possibly be built into the protocol? :^)