syscall_handler: push the interrupt stack instead
make the syscall_handler function to push the interrupt stack instead of
the custom syscall stack (which was basically the interrupt stack but
with the sysret frame instead of the iret frame). This would make things
much easier and in at fork() we can just pop the preserved and scratch
registers and iretq straight away since we have a valid iret frame :^).
This also makes the signal handling much more cleaner since now we dont
have to manage two different frames on the stack :^) too good :^) too
easy :^)

Signed-off-by: Andy-Python-Programmer <andypythonappdeveloper@gmail.com>