x86_64: make use of limine SMP feature
Now that we have switched to the limine boot protocol. We can make use
of the limine SMP feature to boot the APs for us. This means that we can
get rid of our SMP trampoline (this should make build times a bit
faster) and removes unwanted mess from the kernel which should be in the
bootloader (if a bootloader is responsible to boot the BSP then it
should be for the other APs as-well). We also remove the conventional
memory reservation (1st page which was before used to boot up the APs).

Limine :heart:!

Signed-off-by: Andy-Python-Programmer <andypythonappdeveloper@gmail.com>