dma: allocate the nearest size
Currently in the `DmaAllocator`, if the request size was greater than
PAGE_SIZE (4KiB), the allocator would allocate a 2MiB page to satisfy
the allocation request. The memory required to satisfy the request is
relaxed by finding the nearest buddy order from the allocation request
size and allocating with that order instead.

Signed-off-by: Andy-Python-Programmer <andypythonappdeveloper@gmail.com>