VirtAddr::read_mut: return `ReadErr` instead of None
This makes error handling in the syscalls much cleaner. Another thing
that we could do is specifically have a `UserVirtAddr` structure, which
on reads validates that the memory that we are reading is valid for
memory R/W (depending of mutablity) in the read_{mut} function and the
address is a valid userland address (under arch::userland_max_address()).

Signed-off-by: Andy-Python-Programmer <andypythonappdeveloper@gmail.com>