kernel: remove the `syslog` feature for `trace` syscall
```rust
/// Enables syscall tracer for this process/thread.
///
/// When the tracer is enabled for a process, it also applies to any child processes spawned by the
/// process.
pub fn trace() -> Result<usize, SyscallError> { /* ... */ }
```

The trace system call has the same functionality as the old `syslog`
feature, but now it can be enabled per task.

This commit also introduces the `systrace` command which you can run to
get the system call trace for a process. Though currently, the systrace
is printed in the kernel. Maybe in the future we might want to
implement this via `ptrace()` or return the logs to the user?

Signed-off-by: Andy-Python-Programmer <andypythonappdeveloper@gmail.com>