Added missing Ubuntu dependencies > **Note**: This commit message is written in Markdown for the sake of clean documentation.
I'm not entirely sure if `deps_linux` is intended to *only* install
Ubuntu dependencies, as it checks for `apt`, and not the distribution
the script is being run on.

```
if [ -x "$(command -v apt)" ]; then
PKGMAN="apt"
```

Despite this, I'm assuming its intended for Ubuntu, as that seems to
be the primary GNU/Linux distribution used for the compilation of Aero.